+++ routes/index.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const { get } = require('../controllers/index')
-const entriesRouter = require('./news')
+const entriesRouter = require('./entries')
 
 const router = express.Router()
